+++ src/low_level.rs
@@ -59,6 +59,51 @@ where
     }
 }
 
+struct BorrowedSliceVisitor<Enc, T, E>(PhantomData<(Enc, T, E)>);
+
+impl<Enc, T, E> de::Visitor<'_> for BorrowedSliceVisitor<Enc, T, E>
+where
+    Enc: Encoding,
+    T: Clone,
+    for<'a> &'a T: TryFrom<&'a [u8], Error = E>,
+    E: fmt::Display,
+{
+    type Value = T;
+
+    fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "a bytestring")
+    }
+
+    fn visit_str<SE>(self, v: &str) -> Result<Self::Value, SE>
+    where
+        SE: de::Error,
+    {
+        let bytes = Enc::decode(v)?;
+        let bytes_len = bytes.len();
+        let result_ref: &T = AsRef::<[u8]>::as_ref(&bytes).try_into().map_err(|err| {
+            de::Error::custom(format!(
+                "Failed to instantiate `{}` from a byte slice of length {bytes_len}: {err}",
+                type_name::<T>()
+            ))
+        })?;
+        Ok(result_ref.clone())
+    }
+
+    fn visit_bytes<SE>(self, v: &[u8]) -> Result<Self::Value, SE>
+    where
+        SE: de::Error,
+    {
+        let v_len = v.len();
+        let result_ref: &T = v.try_into().map_err(|err| {
+            de::Error::custom(format!(
+                "Failed to instantiate `{}` from a byte slice of length {v_len}: {err}",
+                type_name::<T>()
+            ))
+        })?;
+        Ok(result_ref.clone())
+    }
+}
+
 struct ArrayVisitor<Enc, T, E, const N: usize>(PhantomData<(Enc, T, E)>);
 
 impl<Enc, T, E, const N: usize> de::Visitor<'_> for ArrayVisitor<Enc, T, E, N>
@@ -165,6 +210,22 @@ where
     }
 }
 
+pub(crate) fn deserialize_borrowed_slice<'de, Enc: Encoding, T, E, D>(
+    deserializer: D,
+) -> Result<T, D::Error>
+where
+    D: Deserializer<'de>,
+    T: Clone,
+    for<'a> &'a T: TryFrom<&'a [u8], Error = E>,
+    E: fmt::Display,
+{
+    if deserializer.is_human_readable() {
+        deserializer.deserialize_str(BorrowedSliceVisitor::<Enc, T, E>(PhantomData))
+    } else {
+        deserializer.deserialize_bytes(BorrowedSliceVisitor::<Enc, T, E>(PhantomData))
+    }
+}
+
 pub(crate) fn deserialize_array<'de, Enc: Encoding, const N: usize, T, E, D>(
     deserializer: D,
 ) -> Result<T, D::Error>
@@ -205,7 +266,7 @@ mod tests {
 
     use serde::{Deserialize, Serialize};
 
-    use crate::{encoding::Hex, ArrayLike, SliceLike};
+    use crate::{encoding::Hex, ArrayLike, BorrowedSliceLike, SliceLike};
 
     #[cfg(feature = "generic-array-014")]
     use crate::GenericArray014;
@@ -215,6 +276,9 @@ mod tests {
 
     #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
     struct VectorStruct(#[serde(with = "SliceLike::<Hex>")] Vec<u8>);
+
+    #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
+    struct BorrowStruct<const N: usize>(#[serde(with = "BorrowedSliceLike::<Hex>")] Borrow<N>);
 
     #[cfg(feature = "generic-array-014")]
     #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -228,7 +292,28 @@ mod tests {
     #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
     struct WrongValue(u32);
 
-    #[derive(Debug, PartialEq, Eq)]
+    #[derive(Debug, Clone, PartialEq, Eq)]
+    struct Borrow<const N: usize>([u8; N]);
+
+    impl<const N: usize> AsRef<[u8]> for Borrow<N> {
+        fn as_ref(&self) -> &[u8] {
+            self.0.as_ref()
+        }
+    }
+
+    impl<'a, const N: usize> TryFrom<&'a [u8]> for &'a Borrow<N> {
+        type Error = String;
+        fn try_from(source: &'a [u8]) -> Result<Self, Self::Error> {
+            let source_len = source.len();
+            if source_len != N {
+                return Err(format!("The slice must have length {N}"));
+            }
+
+            Ok(unsafe { &*(source.as_ptr() as *const Borrow<N>) })
+        }
+    }
+
+    #[derive(Debug, Clone, PartialEq, Eq)]
     struct BadType([u8; 4]);
 
     impl AsRef<[u8]> for BadType {
@@ -254,11 +339,24 @@ mod tests {
         }
     }
 
+    impl<'a> TryFrom<&'a [u8]> for &'a BadType {
+        type Error = String;
+        fn try_from(source: &'a [u8]) -> Result<Self, Self::Error> {
+            let source_len = source.len();
+            Err(format!(
+                "BadType cannot deserialize from `&[u8]` of length {source_len}"
+            ))
+        }
+    }
+
     #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
     struct BadArrayStruct(#[serde(with = "ArrayLike::<Hex>")] BadType);
 
     #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
     struct BadSliceStruct(#[serde(with = "SliceLike::<Hex>")] BadType);
+
+    #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
+    struct BadBorrowSliceStruct(#[serde(with = "BorrowedSliceLike::<Hex>")] BadType);
 
     fn bin_serialize<T: Serialize>(value: T) -> Result<Vec<u8>, String> {
         rmp_serde::encode::to_vec(&value).map_err(|err| err.to_string())
@@ -401,6 +499,67 @@ mod tests {
         let bad_struct_bytes = bin_serialize(BadSliceStruct(BadType([1, 2, 3, 4]))).unwrap();
         assert_eq!(
             bin_deserialize::<BadSliceStruct>(&bad_struct_bytes).unwrap_err(),
+            concat![
+                "Failed to instantiate `serde_encoded_bytes::low_level::tests::BadType` ",
+                "from a byte slice of length 4: ",
+                "BadType cannot deserialize from `&[u8]` of length 4"
+            ]
+        );
+    }
+
+    #[test]
+    fn borrow_slice_visitor_human_readable() {
+        let val = BorrowStruct(Borrow([1, 2, 3, 4]));
+
+        // Normal operation
+        let val_str = hr_serialize(&val).unwrap();
+        let val_back = hr_deserialize::<BorrowStruct<4>>(&val_str).unwrap();
+        assert_eq!(val, val_back);
+
+        // Failed to decode
+        assert_eq!(
+            hr_deserialize::<BorrowStruct<4>>("\"0x0102030\"").unwrap_err(),
+            "Odd number of digits at line 1 column 11"
+        );
+
+        // Unexpected value type
+        assert_eq!(
+            hr_deserialize::<BorrowStruct<4>>("1").unwrap_err(),
+            "invalid type: integer `1`, expected a bytestring at line 1 column 1"
+        );
+
+        // A struct that always fails on deserialization
+        let bad_struct_str = hr_serialize(BadBorrowSliceStruct(BadType([1, 2, 3, 4]))).unwrap();
+        assert_eq!(
+            hr_deserialize::<BadBorrowSliceStruct>(&bad_struct_str).unwrap_err(),
+            concat![
+                "Failed to instantiate `serde_encoded_bytes::low_level::tests::BadType` ",
+                "from a byte slice of length 4: ",
+                "BadType cannot deserialize from `&[u8]` of length 4 at line 1 column 12"
+            ]
+        );
+    }
+
+    #[test]
+    fn borrow_slice_visitor_binary() {
+        let val = BorrowStruct(Borrow([1, 2, 3, 4]));
+
+        // Normal operation
+        let val_bytes = bin_serialize(&val).unwrap();
+        let val_back = bin_deserialize::<BorrowStruct<4>>(&val_bytes).unwrap();
+        assert_eq!(val, val_back);
+
+        // Unexpected value type
+        let wrong_val_bytes = bin_serialize(WrongValue(0x01020304)).unwrap();
+        assert_eq!(
+            bin_deserialize::<BorrowStruct<4>>(&wrong_val_bytes).unwrap_err(),
+            "invalid type: integer `16909060`, expected a bytestring"
+        );
+
+        // A struct that always fails on deserialization
+        let bad_struct_bytes = bin_serialize(BadBorrowSliceStruct(BadType([1, 2, 3, 4]))).unwrap();
+        assert_eq!(
+            bin_deserialize::<BadBorrowSliceStruct>(&bad_struct_bytes).unwrap_err(),
             concat![
                 "Failed to instantiate `serde_encoded_bytes::low_level::tests::BadType` ",
                 "from a byte slice of length 4: ",